+++ src/quacc/utils/files.py
@@ -141,13 +141,13 @@ def load_yaml_calc(yaml_path: str | Path) -> dict[str, Any]:
     return config
 
 
-def find_recent_logfile(dir_name: Path | str, logfile_extensions: str | list[str]):
+def find_recent_logfile(directory: Path | str, logfile_extensions: str | list[str]):
     """
     Find the most recent logfile in a given directory.
 
     Parameters
     ----------
-    dir_name
+    directory
         The path to the directory to search
     logfile_extensions
         The extension (or list of possible extensions) of the logfile to search
@@ -162,8 +162,8 @@ def find_recent_logfile(dir_name: Path | str, logfile_extensions: str | list[str
     logfile = None
     if isinstance(logfile_extensions, str):
         logfile_extensions = [logfile_extensions]
-    for f in os.listdir(dir_name):
-        f_path = Path(dir_name, f)
+    for f in os.listdir(directory):
+        f_path = Path(directory, f)
         for ext in logfile_extensions:
             if ext in f and f_path.stat().st_mtime > mod_time:
                 mod_time = f_path.stat().st_mtime
@@ -171,7 +171,7 @@ def find_recent_logfile(dir_name: Path | str, logfile_extensions: str | list[str
     return logfile
 
 
-def get_uri(dir_name: str | Path) -> str:
+def get_uri(directory: str | Path) -> str:
     """
     Return the URI path for a directory.
 
@@ -182,15 +182,15 @@ def get_uri(dir_name: str | Path) -> str:
 
     Parameters
     ----------
-    dir_name : str
+    directory : str
         A directory name.
 
     Returns
     -------
     str
-        Full URI path, e.g., "fileserver.host.com:/full/path/of/dir_name".
+        Full URI path, e.g., "fileserver.host.com:/full/path/of/directory".
     """
-    fullpath = Path(dir_name).absolute()
+    fullpath = Path(directory).absolute()
     hostname = socket.gethostname()
     with contextlib.suppress(socket.gaierror, socket.herror):
         hostname = socket.gethostbyaddr(hostname)[0]