+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- Renamed `dir_name` to `directory` in the schemas.
 - The method for specifying swaps to the default calculator arguments in recipes has been overhauled. In place of the previous `calc_swaps: dict` approach, additional parameters can be specified via the `**kwargs`. In practice, this means in recipes you can do `scf = "diis", max_steps=100` instead of `calc_swaps={"scf":"diis", "max_steps":100}`, for instance.
 
 ## [0.3.13]